pyproject.toml
[tool.poetry]
name = "corrscope"
version = "0.1.1-pre"
description = ""
authors = ["jimbo1qaz <jimbo1qaz@gmail.com>"]
license = "BSD-2-Clause"

[tool.poetry.dependencies]
python = "^3.6"
"ruamel.yaml" = "^0.15.70"
numpy = "^1.15"
click = "^7.0"
more_itertools = "^4.3"
matplotlib = "^3.0"
attrs = "^18.2.0"
PyQt5 = "^5.11"
PyQt5-sip = "^4.19"

[tool.poetry.dev-dependencies]
pytest = "^4.0"
pytest_mock = "^1.10"
hypothesis = "^3.84"
delayed-assert = "^0.2.3"
pyinstaller = "^3.4"
pywin32-ctypes = {version = "^0.2.0",platform = "win32"}

[tool.poetry.scripts]
corr = 'corrscope.cli:main'
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


[tool.black]
line-length = 88
py36 = true
